Add tags to activity
