+++ samples/WebApiSample/appsettings.json
@@ -24,13 +24,7 @@
     }
   },
   "OpenTelemetry": {
-    "Service": "SampleServiceNowRestClient",
-    "Metrics": {
-      "Source": "SampleServiceNowRestClientMetrics"
-    },
-    "Tracing": {
-      "Source": "SampleServiceNowRestClientTracing"
-    }
+    "Service": "SampleServiceNowRestClient"
   },
   "OpenApi": {
     "Title": "ServiceNow Client Web API sample",
