+++ src/Application.Generators/ControllerGenerator.cs
@@ -45,15 +45,13 @@ namespace {namespaces.WebApi}.Controllers
             sourceBuilder.Append($@"
     [ApiController]
     [Route(""{entity.ResourceName}"")]
-    public partial class {entityPascalName}Controller : ControllerBase
+    public partial class {entityPascalName}Controller : RabbidsIncubator.ServiceNowClient.Application.Mvc.ControllerBase
     {{
-        private readonly ILogger _logger;
-
         private readonly I{entityPascalName}Repository _{entityCamelName}Repository;
 
         public {entityPascalName}Controller(ILogger<{entityPascalName}Controller> logger, I{entityPascalName}Repository {entityCamelName}Repository)
+            : base(logger)
         {{
-            _logger = logger;
             _{entityCamelName}Repository = {entityCamelName}Repository;
         }}
 ");
@@ -65,7 +63,7 @@ namespace {namespaces.WebApi}.Controllers
         public async Task<List<{entityPascalName}Model>> Get(int? startIndex, int? limit)
         {{
             var items = await _{entityCamelName}Repository.FindAllAsync(new QueryModel<{entityPascalName}Model>(null, startIndex, limit));
-            _logger.LogDebug(""Number of items found: {{itemsCount}}"", items.Count);
+            ReportListCount(items.Count);
             return items;
         }}
     }}
@@ -79,7 +77,7 @@ namespace {namespaces.WebApi}.Controllers
         public async Task<List<{entityPascalName}Model>> Get([FromQuery] {entityPascalName}Model model, int? startIndex, int? limit)
         {{
             var items = await _{entityCamelName}Repository.FindAllAsync(new QueryModel<{entityPascalName}Model>(model, startIndex, limit));
-            _logger.LogDebug(""Number of items found: {{itemsCount}}"", items.Count);
+            ReportListCount(items.Count);
             return items;
         }}
     }}
