+++ src/Application/Builder/ApplicationBuilderExtensions.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Builder;
+using RabbidsIncubator.ServiceNowClient.Application.Middlewares;
+
+namespace RabbidsIncubator.ServiceNowClient.Application.Builder
+{
+    public static class ApplicationBuilderExtensions
+    {
+        public static IApplicationBuilder UseActivityEnrichment(
+            this IApplicationBuilder builder)
+        {
+            return builder.UseMiddleware<ActivityEnrichmentMiddleware>();
+        }
+    }
+}
